# openapi.json
{
  "components": {
    "schemas": {
...
      "BodyPayload": {
        "properties": {
          "bytes": {
            "items": {
              "type": "integer"
            },
            "type": "array"
          }
        },
        "required": [
          "bytes"
        ],
        "type": "object"
      },
...
      "CommandReceipt": {
        "properties": {
          "cmd_id": {
            "type": "string"
          },
          "dedup_token": {
            "type": "string"
          },
          "rev": {
            "nullable": true,
            "type": "integer"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "cmd_id",
          "dedup_token",
          "status"
        ],
        "type": "object"
      },
...
      "Error": {
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "OkResponse": {
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "type": "object"
      },
...
      "ThreadDetail": {
        "properties": {
          "agent": {
            "type": "string"
          },
          "agentId": {
            "type": "string"
          },
          "archived": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          },
          "focused": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "lastActivity": {
            "type": "string"
          },
          "lastActivityMs": {
            "type": "integer"
          },
          "log": {
            "items": {
              "$ref": "#/components/schemas/ThreadMsg"
            },
            "type": "array"
          },
          "name": {
            "type": "string"
          },
          "paused": {
            "type": "boolean"
          },
          "status": {
            "enum": [
              "MY_TURN",
              "THEIR_TURN",
              "ACTIVE"
            ],
            "type": "string"
          },
          "unread": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "agentId",
          "log"
        ],
        "type": "object"
      },
      "ThreadMsg": {
        "properties": {
          "author": {
            "enum": [
              "user",
              "assistant"
            ],
            "type": "string"
          },
          "auto": {
            "type": "boolean"
          },
          "fileRef": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "tool": {
            "$ref": "#/components/schemas/ToolCall"
          },
          "ts": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "author"
        ],
        "type": "object"
      },
      "ThreadsResponse": {
        "properties": {
          "focusedThreadId": {
            "nullable": true,
            "type": "string"
          },
          "threads": {
            "items": {
              "$ref": "#/components/schemas/ThreadDetail"
            },
            "type": "array"
          }
        },
        "required": [
          "threads"
        ],
        "type": "object"
      },
      "TicketResponse": {
        "properties": {
          "ticket": {
            "type": "string"
          }
        },
        "required": [
          "ticket"
        ],
        "type": "object"
      },
      "ToolCall": {
        "properties": {
          "intent": {
            "type": "string"
          },
          "isError": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "params": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "result": {
            "type": "string"
          },
          "verb": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Context Pilot Orchestrator",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/api/agent/{id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Raw agent view (rev-envelope)",
        "tags": [
          "agent"
        ]
      }
    },
...
    "/api/agent/{id}/avatar": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Delete avatar",
        "tags": [
          "agent"
        ]
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/*": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Avatar image bytes"
          }
        },
        "summary": "Get agent avatar image",
        "tags": [
          "agent"
        ]
      },
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "format": "binary",
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Upload avatar",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/agent/{id}/body/{hash}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "hash",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BodyPayload"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Hydrate content body",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/agent/{id}/command": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandReceipt"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Send command to agent",
        "tags": [
          "agent"
        ]
      }
    },
...
    "/api/agent/{id}/rename": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "name": {
                    "type": "string"
                  }
                },
                "required": [
                  "name"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Rename agent",
        "tags": [
          "agent"
        ]
      }
    },
...
    "/api/agent/{id}/threads": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadsResponse"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Agent threads + conversation",
        "tags": [
          "agent"
        ]
      }
    },
...
    "/api/fleet": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Raw fleet view (rev-envelope)",
        "tags": [
          "fleet"
        ]
      }
    },
...
    "/api/ticket": {
      "post": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponse"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Mint SSE upgrade ticket",
        "tags": [
          "ticket"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:7878"
    }
  ]
}